close tab chrome

+++ command_task/excecutive_command.py
@@ -2,7 +2,9 @@ from application.app import *
 import os
 import webbrowser
 import psutil
-import subprocess
+import re
+from pywinauto import Desktop
+import time
 import re
 
 chrome_path = "C://Program Files//Google//Chrome//Application//chrome.exe"
@@ -20,14 +22,14 @@ def reverse_lookup(dictionary, value):
 
 
 def control_computer(action):
-    if action.lower() == 'turn off':
-        os.system("shutdown /s /t 1")  # Shutdown command
-        return "Turning off computer..."
-    elif action.lower() == 'restart':
-        os.system("shutdown /r /t 1")  # Restart command
-        return "Restarting computer..."
-    else:
-        return "Sorry, I can't perform that action for the computer."
+    # if action.lower() == 'turn off':
+    #     os.system("shutdown /s /t 1")  # Shutdown command
+    #     return "Turning off computer..."
+    # elif action.lower() == 'restart':
+    #     os.system("shutdown /r /t 1")  # Restart command
+    #     return "Restarting computer..."
+    # else:
+    return "Sorry, I can't perform that action for the computer."
 
 
 def manage_applications(action, app_name, search_query=None):
@@ -49,18 +51,22 @@ def manage_applications(action, app_name, search_query=None):
 
     # đóng tab (check xem tiến trình có chạy không trước khi đóng)
     elif action.lower() == 'close':
-        if search_query:
-            if app_name.lower() == 'chrome':
-                if search_query is None:
-                    close_chrome()
-                    return f"Closing Google Chrome successful"
+        if app_name == 'chrome':
+            if search_query:
                 for app_chr in app_chrome:
-                    if app_chr.lower() == search_query.lower():
-                        check_tab = close_chrome_tab(app_chr)
-                        if check_tab == True:
-                            return f"Closing {app_chr} successful"
+                    print(app_chr.lower())
+                    if app_chr.lower() in search_query.lower():
+                        is_run = close_chrome_tab(app_chr.lower())
+                        if is_run:
+                            return f'Close {app_chr} successful'
                         else:
-                            return f"{app_chr} not found or already closed."
+                            return f'{app_chr} is not running'
+            else:
+                is_run = close_chrome()
+                if is_run:
+                    return 'Close Chrome successful'
+                else:
+                    return 'Chrome is not running'
         else:
             # Kiểm tra xem tiến trình có đang chạy hay không
             check_process = os.popen(
@@ -80,46 +86,49 @@ def search_information(query):
     return f"Searching information for: {query}"
 
 
-def find_process_by_name(name):
-    for proc in psutil.process_iter(['pid', 'name']):
-        if proc.info['name'] == name:
-            return proc
-    return None
+def close_chrome_tab(app_chr):
+    pattern = r"^.*?-(.*?)-(.*?)-"
 
+    # Kết nối đến cửa sổ Chrome
+    desktop = Desktop(backend="uia")
+    chrome_window = desktop.window(
+        title_re=".* Google Chrome$", control_type="Pane")
 
-def close_chrome_tab(tab_name):
-    tab_found = False
-    tabs = get_chrome_tabs()
+    # Tìm tất cả các tab trong cửa sổ Chrome
+    tabs = chrome_window.descendants(control_type="TabItem")
+
+    # Xác định tab YouTube và đóng nó
     for tab in tabs:
-        if tab_name.lower() in tab.lower():
-            for process in psutil.process_iter(['pid', 'name', 'cmdline']):
-                if process.info['name'] == 'chrome.exe' and process.info['cmdline']:
-                    cmdline = ' '.join(process.info['cmdline'])
-                    if tab in cmdline:
-                        pid = process.info['pid']
-                        os.system(f'taskkill /F /PID {pid}')
-                        tab_found = True
-
-        # Lấy và in ra danh sách các tab của Chrome hiện tại
-    chrome_tabs = get_chrome_tabs()
-    print("Current Chrome tabs:")
-    for tab in chrome_tabs:
-        print(tab)
-
-    return tab_found
-
-
-def get_chrome_tabs():
-    chrome_tabs = []
-    for process in psutil.process_iter(['pid', 'name', 'cmdline']):
-        if process.info['name'] == 'chrome.exe' and process.info['cmdline']:
-            cmdline = ' '.join(process.info['cmdline'])
-            match = re.search(r'--profile-directory=([\w-]+)', cmdline)
+        # in các tiến trình đang chạy trên chrome
+        # print(tab.window_text)
+        try:
+            match = re.match(pattern, str(tab.window_text))
             if match:
-                profile_directory = match.group(1)
-                chrome_tabs.append(profile_directory)
-    return chrome_tabs
+                if app_chr in match.group(2).lower() or (app_chr in match.group(1).lower() and "usage" in match.group(2).lower()):
+                    chrome_window.set_focus()
+                    tab.click_input(double=True)  # Đảm bảo tab được chọn
+                    time.sleep(1)  # Đợi để đảm bảo tab đã được chọn
+                    tab.type_keys("^w")  # Đóng tab
+                    return True
+
+        except Exception as e:
+            # Tiếp tục vòng lặp
+            print(f"Lỗi: {e}")
+            continue
+
+    return False
+
+
+def is_chrome_running():
+    # Kiểm tra tất cả các tiến trình đang chạy
+    for process in psutil.process_iter(['pid', 'name']):
+        if process.info['name'] == 'chrome.exe':
+            return True
+    return False
 
 
 def close_chrome():
-    os.system("taskkill /IM chrome.exe /F")
+    if is_chrome_running():
+        os.system("taskkill /IM chrome.exe /F")
+        return True
+    return False
